+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/Volume/ProxyVolumeUI.cs
@@ -11,12 +11,12 @@ namespace UnityEditor.Rendering.HighDefinition
         internal static GUIContent shapeContent = EditorGUIUtility.TrTextContent("Shape", "The shape of the Proxy.\nInfinite is compatible with any kind of InfluenceShape.");
         internal static GUIContent boxSizeContent = EditorGUIUtility.TrTextContent("Box Size", "The size of the box.");
         internal static GUIContent sphereRadiusContent = EditorGUIUtility.TrTextContent("Sphere Radius", "The radius of the sphere.");
-		internal static GUIContent planesContent = EditorGUIUtility.TrTextContent("Planes", "The planes to the convex volume.");
-        
-		internal static GUIContent addPlane = EditorGUIUtility.TrTextContent("Add Plane", "Add a new plane to the volume.");
-		internal static GUIContent deleteSelected = EditorGUIUtility.TrTextContent("Delete Selected", "Remove the selected plane from the volume.");
-		internal static GUIContent clearSelection = EditorGUIUtility.TrTextContent("Clear Selection", "Unselect the selected plane.");
-		internal static GUIContent duplicateSelected = EditorGUIUtility.TrTextContent("Duplicate Selected", "Duplicate the selected plane.");
+        internal static GUIContent planesContent = EditorGUIUtility.TrTextContent("Planes", "The planes to the convex volume.");
+
+        internal static GUIContent addPlane = EditorGUIUtility.TrTextContent("Add Plane", "Add a new plane to the volume.");
+        internal static GUIContent deleteSelected = EditorGUIUtility.TrTextContent("Delete Selected", "Remove the selected plane from the volume.");
+        internal static GUIContent clearSelection = EditorGUIUtility.TrTextContent("Clear Selection", "Unselect the selected plane.");
+        internal static GUIContent duplicateSelected = EditorGUIUtility.TrTextContent("Duplicate Selected", "Duplicate the selected plane.");
         internal static GUIContent selectedPlane = EditorGUIUtility.TrTextContent("Selected Plane", "The coefficients of the plane relative to the volume.\nXYZ is the plane normal. W is the distance to origin");
 
         public static readonly CED.IDrawer SectionShape = CED.Group((serialized, owner) =>
@@ -50,7 +50,7 @@ namespace UnityEditor.Rendering.HighDefinition
                     int selected = editing ? serialized.selected.intValue : -1;
                     if (selected >= serialized.planes.arraySize)
                         selected = -1;
-                    
+
                     EditorGUILayout.PropertyField(serialized.planes, planesContent);
                     using (new EditorGUI.DisabledScope(selected == -1))
                     {