+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/SerializedGlobalLightLoopSettings.cs
@@ -26,6 +26,7 @@ namespace UnityEditor.Rendering.HighDefinition
         public SerializedProperty maxEnvLightsOnScreen;
         public SerializedProperty maxDecalsOnScreen;
         public SerializedProperty maxPlanarReflectionOnScreen;
+        public SerializedProperty maxProxyPlanesOnScreen;
 
         public SerializedGlobalLightLoopSettings(SerializedProperty root)
         {
@@ -54,6 +55,7 @@ namespace UnityEditor.Rendering.HighDefinition
             maxEnvLightsOnScreen = root.Find((GlobalLightLoopSettings s) => s.maxEnvLightsOnScreen);
             maxDecalsOnScreen = root.Find((GlobalLightLoopSettings s) => s.maxDecalsOnScreen);
             maxPlanarReflectionOnScreen = root.Find((GlobalLightLoopSettings s) => s.maxPlanarReflectionOnScreen);
+            maxProxyPlanesOnScreen = root.Find((GlobalLightLoopSettings s) => s.maxProxyPlanesOnScreen);
         }
     }
 }