+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightDefinition.cs
@@ -208,6 +208,7 @@ namespace UnityEngine.Rendering.HighDefinition
 
         // Box: extents = box extents
         // Sphere: extents.x = sphere radius
+        // Convex: extents.x = first plane in _ProxyPlaneDatas, extents.y = plane count
         public Vector3 proxyExtents;
         // User can chose if they use This is use in case we want to force infinite projection distance (i.e no projection);
         [SurfaceDataAttributes(precision = FieldPrecision.Real)]