+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -132,6 +132,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added few exposure debug modes.
 - Added support for multiple path-traced views at once (e.g., scene and game views).
 - Added custom target mid grey for auto exposure.
+- Added convex shaped proxies for reflection probes
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)